+++ src/main/java/ladysnake/spawnlanterns/mixin/InfoMixin.java
@@ -6,6 +6,7 @@ import ladysnake.spawnlanterns.common.ExtendedSpawnInfo;
 import ladysnake.spawnlanterns.common.SpawnLanterns;
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.math.ChunkPos;
 import net.minecraft.world.SpawnHelper;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
@@ -31,7 +32,7 @@ public abstract class InfoMixin implements ExtendedSpawnInfo {
     }
 
     @Inject(method = "isBelowCap", at = @At("RETURN"), cancellable = true)
-    private void isBelowCap(SpawnGroup group, CallbackInfoReturnable<Boolean> ci) {
+    private void isBelowCap(SpawnGroup group, ChunkPos chunkPos, CallbackInfoReturnable<Boolean> ci) {
         if (!ci.getReturnValueZ() && (closestPlayer != null && closestPlayer.hasStatusEffect(SpawnLanterns.PROVOCATION))) {
             int i = group.getCapacity() * this.spawningChunkCount / SpawnHelperAccessor.getChunkArea();
             ci.setReturnValue(this.groupToCount.getInt(group) < (i * 2));